[PAGOPA-1446] refactoring
